fix: close W1 contract regressions
