+++ packages/kernel/test/contracts/public-surface.test.ts
@@ -40,7 +40,6 @@ const publicRuntimeSurface = [
   "ProjectionWarningSourceSchema",
   "ProvenanceEntrySchema",
   "PublishableProjectionSchema",
-  "RECOVERY_BUDGET",
   "REPLAY_TASK_GRAPH",
   "RedactionFindingSchema",
   "RelationDirectionSchema",
@@ -62,11 +61,10 @@ const publicRuntimeSurface = [
   "VcsCommandError",
   "VersionControlSystem",
   "VerticalDefinitionSchema",
-  "WriteChainContractError",
+  "WRITE_RECEIPT_SCHEMA",
   "WriteCoordinator",
   "WriteJournalOpSchema",
   "applyTransition",
-  "assertCurrentWriter",
   "assertNoPortablePathCollisions",
   "assertUniqueDocmapIds",
   "auditTaskProvenance",
@@ -83,7 +81,6 @@ const publicRuntimeSurface = [
   "createInMemoryPublishIdempotencyLedger",
   "createTaskIdentity",
   "createWriteReceipt",
-  "decideTaskLifecycleWrite",
   "decisionAmendFieldSupportsOperation",
   "decisionAmendableFields",
   "decisionEntityId",
@@ -137,7 +134,6 @@ const publicRuntimeSurface = [
   "isRuntimeEventKind",
   "isRuntimeEventResultStatus",
   "isTaskWorkKind",
-  "issueWriterGenerationToken",
   "listTaskIndexPaths",
   "makeJournaledWriteCoordinator",
   "makeLocalLockRegistry",
@@ -147,7 +143,6 @@ const publicRuntimeSurface = [
   "makeTaskProjection",
   "moduleEntityId",
   "moduleKeyFromEntityId",
-  "nextRecoveryBatch",
   "normalizeCommandEnvelope",
   "normalizeRelativeDocumentPath",
   "normalizeTaskLifecycleCommand",
@@ -199,8 +194,6 @@ const publicRuntimeSurface = [
   "runtimeEventResultStatuses",
   "scanPrivateText",
   "schemaRegistry",
-  "serializeEventEnvelope",
-  "serializeEventHead",
   "sha256Text",
   "stablePayloadHash",
   "stableStringify",
@@ -212,7 +205,6 @@ const publicRuntimeSurface = [
   "taskPackagePath",
   "taskWorkKinds",
   "unregisterDaemonRepo",
-  "validateActorIdentity",
   "validateExtensionInputShape",
   "validateLifecycleBindingInvariant",
   "validatePresetManifests",
@@ -221,13 +213,10 @@ const publicRuntimeSurface = [
   "validateTaskIdSyntax",
   "validateTemplateCatalog",
   "validateVerticalDefinition",
-  "validateWriteReceipt",
-  "validateWriteSource",
   "warning",
   "writeContentAddressedBlob",
   "writeCoordinatedPayload",
   "writeCoordinatedTaskDocuments",
-  "writeReceiptOutcomes",
   "writeSessionEntity"
 ];
