+++ tools/gates/line-budgets.json
@@ -6,7 +6,7 @@
     "write-contract": 350,
     "doc-sync": 679,
     "preset": 6926,
-    "cli": 16721,
+    "cli": 16717,
     "gui": 15309,
     "daemon": 2333,
     "authority-write-path": 0,
